# Optimize copy_on_write.ConvertFileToCOW