This method is used when unpacking container images and when saving the initial memory snapshot. The latter is on the critical path of the firecracker `pause` stage.

By using io.Copy with 2 files, we allow it to use the `copy_file_range` system call. This avoids copying memory between kernel and user space. Then we don't need a buffer in Go, which also allows copying bigger ranges instead of just one block at a time. This is where most of the improvement comes from.

Passing chunk starts through a channel instead of starting a new goroutine for every chunk gives another 1-2% improvement.

All together, this is about 36% faster on 100MB files without holes. Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor
                               │     base      │              optimized               │
                               │    sec/op     │    sec/op     vs base                │
ConvertFileToCOW/0Sparse-24      12.532m ±  1%   7.959m ±  3%  -36.49% (p=0.000 n=15)
ConvertFileToCOW/0.1Sparse-24    12.100m ±  0%   7.903m ±  1%  -34.68% (p=0.000 n=15)
ConvertFileToCOW/0.5Sparse-24     7.905m ±  0%   5.630m ±  1%  -28.78% (p=0.000 n=15)
ConvertFileToCOW/0.9Sparse-24     1.994m ±  1%   1.512m ±  1%  -24.17% (p=0.000 n=15)
ConvertFileToCOW/0.99Sparse-24    417.0µ ±  2%   382.9µ ±  6%   -8.16% (p=0.000 n=15)
ConvertFileToCOW/1Sparse-24       108.8µ ± 14%   123.7µ ± 26%        ~ (p=0.512 n=15)
geomean                           2.184m         1.714m        -21.51%

                               │     base     │              optimized              │
                               │     B/op     │     B/op      vs base               │
ConvertFileToCOW/0Sparse-24      3.962Mi ± 0%   3.925Mi ± 0%  -0.93% (p=0.000 n=15)
ConvertFileToCOW/0.1Sparse-24    3.963Mi ± 0%   3.925Mi ± 0%  -0.94% (p=0.000 n=15)
ConvertFileToCOW/0.5Sparse-24    3.964Mi ± 0%   3.925Mi ± 0%  -0.97% (p=0.000 n=15)
ConvertFileToCOW/0.9Sparse-24    3.966Mi ± 0%   3.925Mi ± 0%  -1.02% (p=0.000 n=15)
ConvertFileToCOW/0.99Sparse-24   3.967Mi ± 0%   3.924Mi ± 0%  -1.10% (p=0.000 n=15)
ConvertFileToCOW/1Sparse-24      3.915Mi ± 0%   3.914Mi ± 0%  -0.03% (p=0.000 n=15)
geomean                          3.956Mi        3.923Mi       -0.83%

                               │    base     │             optimized              │
                               │  allocs/op  │ allocs/op   vs base                │
ConvertFileToCOW/0Sparse-24       322.0 ± 0%   290.0 ± 0%   -9.94% (p=0.000 n=15)
ConvertFileToCOW/0.1Sparse-24     324.0 ± 0%   293.0 ± 0%   -9.57% (p=0.000 n=15)
ConvertFileToCOW/0.5Sparse-24     325.0 ± 0%   293.0 ± 0%   -9.85% (p=0.000 n=15)
ConvertFileToCOW/0.9Sparse-24     326.0 ± 0%   293.0 ± 0%  -10.12% (p=0.000 n=15)
ConvertFileToCOW/0.99Sparse-24    305.0 ± 0%   265.0 ± 0%  -13.11% (p=0.000 n=15)
ConvertFileToCOW/1Sparse-24      128.00 ± 0%   90.00 ± 1%  -29.69% (p=0.000 n=15)
geomean                           274.9        236.3       -14.05%
```